u32 -> i32 and f32 -> f64

+++ src/models/grade.rs
@@ -5,18 +5,18 @@ use std::hash::{Hash, Hasher};
 pub struct Grade {
     pub course: String,
     pub class: String,
-    pub year: u32,
+    pub year: i32,
     pub name: String,
-    pub class_average: f32,
+    pub class_average: f64,
 }
 
 impl Grade {
     pub fn new(
         course: String,
         class: String,
-        year: u32,
+        year: i32,
         name: String,
-        class_average: f32,
+        class_average: f64,
     ) -> Grade {
         Grade {
             course,
